add property of rwd500 full screen and use blueprint at app.py
