+++ app.py
@@ -1,19 +1,25 @@
 from flask import *
 from flask import jsonify
 
+#blueprint引入--------------------------------------------
+from view.attractions import attractions_blueprint
+from view.attraction_id import attraction_id_blueprint
+from view.categories import categories_blueprint
+
+
 app=Flask(__name__)
 app.config["JSON_AS_ASCII"]=False
 app.config["TEMPLATES_AUTO_RELOAD"]=True
 
-import mysql.connector
-mydb=mysql.connector.connect(
-    host="localhost",
-    user="root",
-    password="Bb0970662139"
-)
-mycursor=mydb.cursor()
-sql="USE taipei_day_trip"
-mycursor.execute(sql)
+# import mysql.connector
+# mydb=mysql.connector.connect(
+#     host="localhost",
+#     user="root",
+#     password="Bb0970662139"
+# )
+# mycursor=mydb.cursor()
+# sql="USE taipei_day_trip"
+# mycursor.execute(sql)
 
 
 # Pages
@@ -33,179 +39,180 @@ def thankyou():
 
 
 
-
-@app.route("/api/attractions",methods=["GET"])
-def api_attractions():
-	# try:
+app.register_blueprint(attractions_blueprint)
+# @app.route("/api/attractions",methods=["GET"])
+# def api_attractions():
+# 	# try:
 	
-	page_num=int(request.args.get("page",""))
-	keyword=request.args.get("keyword","")
-	page_maxnum=12
-	##單純頁數
-	if keyword == "":
-		mycursor=mydb.cursor()
-		sql_fetchall="SELECT *from datas3 LIMIT %s,%s"
-		first_num=page_num*0
-		last_num=(page_num+1)*page_maxnum-1
-		adr_fetchall=(page_num*page_maxnum,page_maxnum)
-		mycursor.execute(sql_fetchall,adr_fetchall)
-		myresult_fetchall=mycursor.fetchall()
-		total_amount=len(myresult_fetchall)
-		# print(myresult_fetchall[0])
-		# print(myresult_fetchall)       
+# 	page_num=int(request.args.get("page",""))
+# 	keyword=request.args.get("keyword","")
+# 	page_maxnum=12
+# 	##單純頁數
+# 	if keyword == "":
+# 		mycursor=mydb.cursor()
+# 		sql_fetchall="SELECT *from datas3 LIMIT %s,%s"
+# 		first_num=page_num*0
+# 		last_num=(page_num+1)*page_maxnum-1
+# 		adr_fetchall=(page_num*page_maxnum,page_maxnum)
+# 		mycursor.execute(sql_fetchall,adr_fetchall)
+# 		myresult_fetchall=mycursor.fetchall()
+# 		total_amount=len(myresult_fetchall)
+# 		# print(myresult_fetchall[0])
+# 		# print(myresult_fetchall)       
 
-		for i in range(len(myresult_fetchall)):
-			myresult_fetchall[i]=list(myresult_fetchall[i])
-			# print()
-			# print(myresult_fetchall[i][9])
-			myresult_fetchall[i][9]=myresult_fetchall[i][9].split(" ")
+# 		for i in range(len(myresult_fetchall)):
+# 			myresult_fetchall[i]=list(myresult_fetchall[i])
+# 			# print()
+# 			# print(myresult_fetchall[i][9])
+# 			myresult_fetchall[i][9]=myresult_fetchall[i][9].split(" ")
 
-		if total_amount !=0:
-			mytitle = mycursor.description
-			column_name =[col[0] for col in mytitle]
-			data=[]
-			for i in range(total_amount):
-				data.append(dict(zip(column_name,myresult_fetchall[i])))
-			if len(data)==page_maxnum:
-				next_page=page_num+1
-			else:
-				next_page=None
-		else:
-			return jsonify({                              
-					"error": True,
-					"message": "沒有資料了"	
-				})
+# 		if total_amount !=0:
+# 			mytitle = mycursor.description
+# 			column_name =[col[0] for col in mytitle]
+# 			data=[]
+# 			for i in range(total_amount):
+# 				data.append(dict(zip(column_name,myresult_fetchall[i])))
+# 			if len(data)==page_maxnum:
+# 				next_page=page_num+1
+# 			else:
+# 				next_page=None
+# 		else:
+# 			return jsonify({                              
+# 					"error": True,
+# 					"message": "沒有資料了"	
+# 				})
 
-		return jsonify({                              
-					"nextPage": next_page,
-					"data": data	
-				})
-
-
+# 		return jsonify({                              
+# 					"nextPage": next_page,
+# 					"data": data	
+# 				})
 
 
-	#模糊&完全
-	else:
-		mycursor=mydb.cursor()
-		sql_keyword="SELECT *FROM datas3 WHERE category=%s or name like concat('%',%s,'%') LIMIT %s,%s"
-		first_num=page_num*0
-		last_num=(page_num+1)*page_maxnum-1
-		adr_keyword=(keyword,keyword,page_num*page_maxnum,page_maxnum)
-		mycursor.execute(sql_keyword,adr_keyword)
-		myresult_keyword=mycursor.fetchall()
-		total_amount=len(myresult_keyword)
-
-		for i in range(len(myresult_keyword)):
-			myresult_keyword[i]=list(myresult_keyword[i])
-			myresult_keyword[i][9]=myresult_keyword[i][9].split(" ")
-
-		if total_amount !=0:
-			mytitle = mycursor.description
-			column_name =[col[0] for col in mytitle]
-		# 	print()
-		# 	print(column_name)
-		# 	print(myresult_keyword)
-			data=[]
-			for i in range(total_amount):
-				data.append(dict(zip(column_name,myresult_keyword[i])))
-			if len(data)==page_maxnum:
-				next_page=page_num+1
-			else:
-				next_page=None
-		else:
-			return jsonify({                              
-					"error": True,
-					"message": "沒有資料了"	
-				})
-
-		return jsonify({                              
-					"nextPage": next_page,
-					"data": data	
-				})
-	# except:
-	# 	return jsonify({
-	# 		"error":True,
-	# 		"message":"奇奇妙妙錯誤"
-	# 	})
-
-	# finally:
-	# 	mydb.close()
 
 
-@app.route("/api/attraction/<attractionId>",methods=["GET"])
-def api_attractionId(attractionId):
-	get_attractionId=attractionId
-	# print(content)
-	mycursor=mydb.cursor()
-	sql="SELECT *FROM datas3 WHERE id=%s"
-	adr=(get_attractionId,)
-	mycursor.execute(sql,adr)
-	myresult=mycursor.fetchall()
-	# print(myresult)
-	if myresult != []:
-		mytitle = mycursor.description
-		column_name =[col[0] for col in mytitle]
-		# print(column_name)
-		# print(myresult[0])
-		data=[]
-		for i in range(0,1):
-			data.append(dict(zip(column_name,list(myresult[i]))))
-		data=data[0]
-		data["images"]=data["images"].split(" ")
-		imgs_notnull=[]
-		for e in data["images"]:
-			if e!="" and ".MP3" not in e and ".FLV" not in e:
-				imgs_notnull.append(e)
-		data["images"]=imgs_notnull
-		print(data)
+# 	#模糊&完全
+# 	else:
+# 		mycursor=mydb.cursor()
+# 		sql_keyword="SELECT *FROM datas3 WHERE category=%s or name like concat('%',%s,'%') LIMIT %s,%s"
+# 		first_num=page_num*0
+# 		last_num=(page_num+1)*page_maxnum-1
+# 		adr_keyword=(keyword,keyword,page_num*page_maxnum,page_maxnum)
+# 		mycursor.execute(sql_keyword,adr_keyword)
+# 		myresult_keyword=mycursor.fetchall()
+# 		total_amount=len(myresult_keyword)
 
-		return jsonify({
-				"data": data
-			})
+# 		for i in range(len(myresult_keyword)):
+# 			myresult_keyword[i]=list(myresult_keyword[i])
+# 			myresult_keyword[i][9]=myresult_keyword[i][9].split(" ")
 
-	else:
-		return jsonify({
-				"error": True,
-				"message":"景點編號不正確"
-			})
+# 		if total_amount !=0:
+# 			mytitle = mycursor.description
+# 			column_name =[col[0] for col in mytitle]
+# 		# 	print()
+# 		# 	print(column_name)
+# 		# 	print(myresult_keyword)
+# 			data=[]
+# 			for i in range(total_amount):
+# 				data.append(dict(zip(column_name,myresult_keyword[i])))
+# 			if len(data)==page_maxnum:
+# 				next_page=page_num+1
+# 			else:
+# 				next_page=None
+# 		else:
+# 			return jsonify({                              
+# 					"error": True,
+# 					"message": "沒有資料了"	
+# 				})
+
+# 		return jsonify({                              
+# 					"nextPage": next_page,
+# 					"data": data	
+# 				})
+# 	# except:
+# 	# 	return jsonify({
+# 	# 		"error":True,
+# 	# 		"message":"奇奇妙妙錯誤"
+# 	# 	})
+
+# 	# finally:
+# 	# 	mydb.close()
 
 
-@app.route("/api/categories",methods=["GET"])
-def api_cotegories():
-	try:	
-		category_data=[]
-		mycursor2=mydb.cursor()
-		sql2="SELECT DISTINCT category FROM datas3"
-		mycursor2.execute(sql2)
-		myresult_category=mycursor2.fetchall()
-		# for e in myresult_category:
-		# 	category_data.append(str(e).replace("(","").replace(")","").replace(",","").replace("'",""))
-		category_data = [cat[0] for cat in myresult_category]
+app.register_blueprint(attraction_id_blueprint)
+# @app.route("/api/attraction/<attractionId>",methods=["GET"])
+# def api_attractionId(attractionId):
+# 	get_attractionId=attractionId
+# 	# print(content)
+# 	mycursor=mydb.cursor()
+# 	sql="SELECT *FROM datas3 WHERE id=%s"
+# 	adr=(get_attractionId,)
+# 	mycursor.execute(sql,adr)
+# 	myresult=mycursor.fetchall()
+# 	# print(myresult)
+# 	if myresult != []:
+# 		mytitle = mycursor.description
+# 		column_name =[col[0] for col in mytitle]
+# 		# print(column_name)
+# 		# print(myresult[0])
+# 		data=[]
+# 		for i in range(0,1):
+# 			data.append(dict(zip(column_name,list(myresult[i]))))
+# 		data=data[0]
+# 		data["images"]=data["images"].split(" ")
+# 		imgs_notnull=[]
+# 		for e in data["images"]:
+# 			if e!="" and ".MP3" not in e and ".FLV" not in e:
+# 				imgs_notnull.append(e)
+# 		data["images"]=imgs_notnull
+# 		print(data)
+
+# 		return jsonify({
+# 				"data": data
+# 			})
+
+# 	else:
+# 		return jsonify({
+# 				"error": True,
+# 				"message":"景點編號不正確"
+# 			})
+#-------------------------------------------------#-------------------------------------------------
+app.register_blueprint(categories_blueprint)
+# @app.route("/api/categories",methods=["GET"])
+# def api_cotegories():
+# 	try:	
+# 		category_data=[]
+# 		mycursor2=mydb.cursor()
+# 		sql2="SELECT DISTINCT category FROM datas3"
+# 		mycursor2.execute(sql2)
+# 		myresult_category=mycursor2.fetchall()
+# 		# for e in myresult_category:
+# 		# 	category_data.append(str(e).replace("(","").replace(")","").replace(",","").replace("'",""))
+# 		category_data = [cat[0] for cat in myresult_category]
 		
-		# print(myresult_category)
-		print(category_data)
-		return jsonify({
-					"data": category_data
+# 		# print(myresult_category)
+# 		print(category_data)
+# 		return jsonify({
+# 					"data": category_data
 					
-				})
+# 				})
 
-	except TypeError:
-		return jsonify({
-					"error": True,
-					"message":"型別發生錯誤"
-				})
+# 	except TypeError:
+# 		return jsonify({
+# 					"error": True,
+# 					"message":"型別發生錯誤"
+# 				})
 
-	except NameError:
-		return jsonify({
-					"error": True,
-					"message":"使用沒有被定義的對象"
-				})
+# 	except NameError:
+# 		return jsonify({
+# 					"error": True,
+# 					"message":"使用沒有被定義的對象"
+# 				})
 				
-	except Exception:
-		return jsonify({
-					"error": True,
-					"message":"不知道怎麼了，反正發生錯誤惹"
-				})
+# 	except Exception:
+# 		return jsonify({
+# 					"error": True,
+# 					"message":"不知道怎麼了，反正發生錯誤惹"
+# 				})
