+++ static/css/attraction.css
@@ -30,9 +30,11 @@ body nav div {
     width: 85%;
   }
 }
-@media (max-width: 360px) {
+@media (max-width: 500px) {
   body nav div {
     width: 100%;
+    padding-left: 5px;
+    padding-right: 5px;
   }
 }
 body nav div div.leftBar {
@@ -101,6 +103,11 @@ body main .sectionPicture {
     display: block;
   }
 }
+@media (max-width: 500px) {
+  body main .sectionPicture {
+    width: 100%;
+  }
+}
 @media (max-width: 360px) {
   body main .sectionPicture {
     width: 100%;
@@ -111,6 +118,7 @@ body main .sectionPicture .pictureLeft {
   height: 406px;
   margin-top: 37px;
   margin-left: 5px;
+  background-color: rgb(227, 190, 141);
   position: relative;
 }
 @media (max-width: 750px) {
