+++ static/css/index.css
@@ -67,8 +67,9 @@ main .page-one {
   width: 100%;
   height: 320px;
   background-image: url(../img/welcome.png);
-  background-repeat: no-repeat;
   background-size: cover;
+  -o-object-position: 100% 0;
+     object-position: 100% 0;
 }
 main .page-one .text {
   position: absolute;
